refactor : 저장소 기능을 추상화
