+++ src/test/java/chess/domain/MemoryRepositoryTest.java
@@ -1,9 +1,10 @@
 package chess.domain;
 
-import static chess.domain.ChessGame.START_COLOR;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.board.MemoryChessBoard;
+import chess.domain.board.ChessBoard;
+import chess.domain.board.DefaultBoardInitializer;
+import chess.domain.board.MemoryRepository;
 import chess.domain.position.Column;
 import chess.domain.position.Position;
 import chess.domain.position.Row;
@@ -12,16 +13,17 @@ import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-class MemoryChessBoardTest {
+class MemoryRepositoryTest {
 
     @DisplayName("보드 생성기로 보드 생성 시 32개의 기물이 초기화된다.")
     @Test
     void printMap() {
-        MemoryChessBoard memoryChessBoard = new MemoryChessBoard(new HashMap<>());
+        MemoryRepository memoryRepository = new MemoryRepository(new HashMap<>());
+        ChessBoard chessBoard = new ChessBoard(memoryRepository);
 
-        memoryChessBoard.initNewBoard(START_COLOR);
-        Map<Position, Piece> boardMap = memoryChessBoard.getBoard();
         Map<Position, Piece> defaultMap = new HashMap<>();
+        Map<Position, Piece> boardMap = memoryRepository.getBoard();
+        chessBoard.initNewBoard(DefaultBoardInitializer.initializer());
 
         defaultMap.put(new Position(Row.RANK1, Column.A), new Piece(PieceType.ROOK, Color.WHITE));
         defaultMap.put(new Position(Row.RANK1, Column.H), new Piece(PieceType.ROOK, Color.WHITE));
@@ -58,55 +60,5 @@ class MemoryChessBoardTest {
         defaultMap.put(new Position(Row.RANK7, Column.E), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
 
         assertThat(boardMap).isEqualTo(defaultMap);
-    }
-
-    @DisplayName("체스판에 왕이 하나만 있는 경우 false 를 리턴한다.")
-    @Test
-    void hasTwoKingFalseTest() {
-        Map<Position, Piece> boardMap = new HashMap<>();
-        MemoryChessBoard memoryChessBoard = new MemoryChessBoard(boardMap);
-
-        boardMap.put(new Position(Row.RANK8, Column.B), new Piece(PieceType.KING, Color.BLACK));
-        boardMap.put(new Position(Row.RANK8, Column.C), new Piece(PieceType.ROOK, Color.BLACK));
-        boardMap.put(new Position(Row.RANK7, Column.A), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
-        boardMap.put(new Position(Row.RANK7, Column.C), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
-        boardMap.put(new Position(Row.RANK7, Column.D), new Piece(PieceType.BISHOP, Color.BLACK));
-        boardMap.put(new Position(Row.RANK6, Column.B), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
-        boardMap.put(new Position(Row.RANK6, Column.E), new Piece(PieceType.QUEEN, Color.BLACK));
-
-        boardMap.put(new Position(Row.RANK4, Column.F), new Piece(PieceType.KNIGHT, Color.WHITE));
-        boardMap.put(new Position(Row.RANK4, Column.G), new Piece(PieceType.QUEEN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK3, Column.F), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK3, Column.H), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK2, Column.F), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK2, Column.G), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK1, Column.E), new Piece(PieceType.ROOK, Color.WHITE));
-
-        assertThat(memoryChessBoard.hasTwoKing()).isFalse();
-    }
-
-    @DisplayName("체스판에 왕이 모두 살아있는 경우 true 를 리턴한다.")
-    @Test
-    void hasTwoKingTrueTest() {
-        Map<Position, Piece> boardMap = new HashMap<>();
-        MemoryChessBoard memoryChessBoard = new MemoryChessBoard(boardMap);
-
-        boardMap.put(new Position(Row.RANK8, Column.B), new Piece(PieceType.KING, Color.BLACK));
-        boardMap.put(new Position(Row.RANK8, Column.C), new Piece(PieceType.ROOK, Color.BLACK));
-        boardMap.put(new Position(Row.RANK7, Column.A), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
-        boardMap.put(new Position(Row.RANK7, Column.C), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
-        boardMap.put(new Position(Row.RANK7, Column.D), new Piece(PieceType.BISHOP, Color.BLACK));
-        boardMap.put(new Position(Row.RANK6, Column.B), new Piece(PieceType.BLACK_PAWN, Color.BLACK));
-        boardMap.put(new Position(Row.RANK6, Column.E), new Piece(PieceType.QUEEN, Color.BLACK));
-
-        boardMap.put(new Position(Row.RANK4, Column.F), new Piece(PieceType.KNIGHT, Color.WHITE));
-        boardMap.put(new Position(Row.RANK4, Column.G), new Piece(PieceType.QUEEN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK3, Column.F), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK3, Column.H), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK2, Column.F), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK2, Column.G), new Piece(PieceType.WHITE_PAWN, Color.WHITE));
-        boardMap.put(new Position(Row.RANK1, Column.E), new Piece(PieceType.KING, Color.WHITE));
-
-        assertThat(memoryChessBoard.hasTwoKing()).isTrue();
     }
 }
