+++ src/test/java/chess/domain/chessGame/BishopTest.java
@@ -2,7 +2,8 @@ package chess.domain.chessGame;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.domain.board.MemoryChessBoard;
+import chess.domain.board.ChessBoard;
+import chess.domain.board.MemoryRepository;
 import chess.domain.ChessGame;
 import chess.domain.Color;
 import chess.domain.position.Column;
@@ -33,13 +34,13 @@ class BishopTest {
     void generateMovablePositions() {
         Position targetPosition = new Position(Row.RANK1, Column.C);
         Color currentTurn = Color.WHITE;
-        ChessGame chessGame = new ChessGame(new MemoryChessBoard(
+        ChessGame chessGame = new ChessGame(new ChessBoard(new MemoryRepository(
                 Map.of(
                         targetPosition, new Piece(PieceType.BISHOP, currentTurn),
                         new Position(Row.RANK2, Column.B), new Piece(PieceType.WHITE_PAWN, currentTurn),
                         new Position(Row.RANK3, Column.E), new Piece(PieceType.BLACK_PAWN, currentTurn.opposite())
                 )
-        ), currentTurn);
+        )), currentTurn);
 
         List<Position> result = chessGame.generateMovablePositions(targetPosition);
